Service to access endpoints by rol

+++ app/Http/Controllers/ProjectController.php
@@ -3,11 +3,20 @@
 namespace App\Http\Controllers;
 
 use App\Models\Project;
+use App\Providers\JwtServiceProvider;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
 class ProjectController extends Controller
 {
+
+    protected $jwtService;
+
+    public function __construct(JwtServiceProvider $jwtService)
+    {
+        $this->jwtService = $jwtService;
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -22,25 +31,29 @@ class ProjectController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = Validator::make($request->all(), [
-            'c_name' => 'required|string|min:3|max:255',
-            'fk_user' => 'required|numeric|exists:users,pk_user'
-        ], [
-            'c_name.required' => 'El nombre es requerido.',
-            'c_name.string' => 'El nombre debe ser un texto.',
-            'c_name.min' => 'El nombre debe tener mínimo 3 carácteres.',
-            'c_name.max' => 'El nombre debe tener máximo 255 carácteres.',
-            'fk_user.required' => 'El usuario es requerido.',
-            'fk_user.numeric' => 'El usuario debe ser un número.',
-            'fk_user.exists' => 'El usuario no existe.'
-        ]);
+        if ($this->jwtService->haveAccess($request)) {
+            $validator = Validator::make($request->all(), [
+                'c_name' => 'required|string|min:3|max:255',
+                'fk_user' => 'required|numeric|exists:users,pk_user'
+            ], [
+                'c_name.required' => 'El nombre es requerido.',
+                'c_name.string' => 'El nombre debe ser un texto.',
+                'c_name.min' => 'El nombre debe tener mínimo 3 carácteres.',
+                'c_name.max' => 'El nombre debe tener máximo 255 carácteres.',
+                'fk_user.required' => 'El usuario es requerido.',
+                'fk_user.numeric' => 'El usuario debe ser un número.',
+                'fk_user.exists' => 'El usuario no existe.'
+            ]);
 
-        if ($validator->fails()) return response()->json(['message' => $validator->errors()], 400);
+            if ($validator->fails()) return response()->json(['message' => $validator->errors()], 400);
 
-        $project = Project::create($request->all());
-        $project->load('user');
+            $project = Project::create($request->all());
+            $project->load('user');
 
-        return response()->json(['data' => $project, 'message' => 'Proyecto creado.'], 201);
+            return response()->json(['data' => $project, 'message' => 'Proyecto creado.'], 201);
+        } else {
+            return response()->json(['message' => 'No tienes acceso.'], 401);
+        }
     }
 
     /**
@@ -60,38 +73,46 @@ class ProjectController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $validator = Validator::make($request->all(), [
-            'c_name' => 'required|string|min:3|max:255'
-        ], [
-            'c_name.required' => 'El nombre es requerido.',
-            'c_name.string' => 'El nombre debe ser un texto.',
-            'c_name.min' => 'El nombre debe tener mínimo 3 carácteres.',
-            'c_name.max' => 'El nombre debe tener máximo 255 carácteres.'
-        ]);
+        if ($this->jwtService->haveAccess($request)) {
+            $validator = Validator::make($request->all(), [
+                'c_name' => 'required|string|min:3|max:255'
+            ], [
+                'c_name.required' => 'El nombre es requerido.',
+                'c_name.string' => 'El nombre debe ser un texto.',
+                'c_name.min' => 'El nombre debe tener mínimo 3 carácteres.',
+                'c_name.max' => 'El nombre debe tener máximo 255 carácteres.'
+            ]);
 
-        if ($validator->fails()) return response()->json(['message' => $validator->errors()], 400);
+            if ($validator->fails()) return response()->json(['message' => $validator->errors()], 400);
 
-        $project = Project::find($id);
-        if (!$project) return response()->json(['data' => null, 'message' => 'El proyecto no existe.'], 404);
+            $project = Project::find($id);
+            if (!$project) return response()->json(['data' => null, 'message' => 'El proyecto no existe.'], 404);
 
-        $project->c_name = $request->c_name;
-        $project->update();
+            $project->c_name = $request->c_name;
+            $project->update();
 
-        return response()->json(['data' => $project]);
+            return response()->json(['data' => $project]);
+        } else {
+            return response()->json(['message' => 'No tienes acceso.'], 401);
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, string $id)
     {
-        $project = Project::find($id);
-        $project->load('tasks');
-        if (!$project) return response()->json(['data' => null, 'message' => 'El proyecto no existe.'], 404);
+        if ($this->jwtService->haveAccess($request)) {
+            $project = Project::find($id);
+            $project->load('tasks');
+            if (!$project) return response()->json(['data' => null, 'message' => 'El proyecto no existe.'], 404);
 
-        if ($project->tasks && count($project->tasks) > 0) return response()->json(['data' => null, 'message' => 'El proyecto no se puede eliminar ya que tiene tareas asignadas.'], 403);
+            if ($project->tasks && count($project->tasks) > 0) return response()->json(['data' => null, 'message' => 'El proyecto no se puede eliminar ya que tiene tareas asignadas.'], 403);
 
-        $project->delete();
-        return response()->json(['data' => $project, 'message' => 'Proyecto eliminado.']);
+            $project->delete();
+            return response()->json(['data' => $project, 'message' => 'Proyecto eliminado.']);
+        } else {
+            return response()->json(['message' => 'No tienes acceso.'], 401);
+        }
     }
 }
